[x86] Move src/port.h to src/arch/x86 #70
