+++ src/pit.c
@@ -2,7 +2,8 @@
 #include "kinttypes.h"
 #include "panic.h"
 #include "pit.h"
-#include "port.h"
+
+#include "arch/x86/port.h"
 
 #define PORT_CMD      0x0043
 #define PORT_CH0_DATA 0x0040
