+++ src/vga.c
@@ -10,8 +10,9 @@
 #include "arch_vmm.h"
 #include "assert.h"
 #include "memfun.h"
-#include "port.h"
 #include "vga.h"
+
+#include "arch/x86/port.h"
 
 #define VGA_NUM_ROWS 25
 #define VGA_NUM_COLS 80
